Walk back changing radt instance validation to be more parser-like
Maybe this still makes sense for the byte parsing, but it seemed to be
more trouble than it was worth for the structured validation.
In particular, there was no good way to represent empty input with how I
was doing it. I think the structured nature of the input means it makes
more sense as a standard recursive function