First pass at implementing a bridged type.
API is a bit weird, but I think it's what's necessary.
When you encode a bridged type, you get it as a TypedValue,
and a list of other Items to store.
When you decode one, you need to proactively provide all its
dependencies in a map. This will be done by using the "fetching
strategy" for the type, although those aren't implemented yet.